back/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
cffi==2.0.0
click==8.3.0
cryptography==46.0.3
ecdsa==0.19.1
fastapi==0.121.2
greenlet==3.2.4
h11==0.16.0
idna==3.11
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.5.0
pytz==2024.2
rsa==4.9.1
setuptools==70.3.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlalchemy-serializer==1.4.22
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
